+++ Changes.md
@@ -53,7 +53,9 @@ Improvements
 - HierarchyView :
   - Added support for inclusion and exclusion of leaf level locations to the Visible Set.
   - Added <kbd>F</kbd> shortcut to expand ancestors and frame selected paths. This replaces the previous behaviour of automatic updates to expansion and framing in the HierarchyView on selection change.
-- CyclesOptions : Added `useLightTree` plug.
+- CyclesOptions :
+  - Added `useLightTree` plug.
+  - Added path guiding options.
 - CyclesAttributes :
   - Added `emissionSamplingMethod` plug.
   - Added `isCausticsCaster` and `isCausticsReceiver` plugs.